api/requirements.txt
alembic==1.13.1
asyncpg==0.29.0
databases==0.9.0
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg-pool==3.1.6
fastapi[all]==0.110.0
orjson==3.9.15
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
uvicorn==0.28.0
sqlalchemy==2.0.28
python-jose==3.3.0
passlib==1.7.4
redis==5.0.3
bcrypt==4.1.2
python-multipart==0.0.9
pytest
httpx==0.26.0
redis==5.0.3
datamodel-code-generator==0.25.1
pillow==10.2.0
flake8==7.0.0